⚡️ :: dto 추가

+++ src/main/kotlin/com/eathub/eathub/domain/food/application/presentation/dto/FoodApplicationMessage.kt
@@ -5,7 +5,7 @@ class FoodApplicationMessages(
 )
 
 class FoodApplicationRestaurantMessages(
-    val countSum: Long,
+    val countSum: Int,
     val costSum: Long,
     val restaurantName: String,
     val applications: List<FoodApplicationMessage>
